refactor(services): shrink evidence-rescore function; split JobManager monitor behavior into mixin
